Clean up Marker, remove dependency on libunwind
