+++ nativelib/src/main/resources/gc/immix/Log.h
@@ -4,6 +4,7 @@
 //#define NDEBUG
 #include <assert.h>
 
+//#define PRINT_STACK_OVERFLOW
 //#define DEBUG_PRINT
 
 #endif // IMMIX_LOG_H
